+++ client/src/core.js
@@ -7,7 +7,7 @@ import '@/style'
 import $ from 'jquery'
 import 'jquery-migrate'
 
-import { initialize, ajreq } from '@/script'
+import { initialize } from '@/script'
 
 import 'jquery-ui'
 import registerServiceWorker from '@/registerServiceWorker'
@@ -32,32 +32,6 @@ import { getCsrfToken } from '@/api/base'
 
 initialize()
 registerServiceWorker()
-
-$('#fs-profile-rate-comment').dialog({
-  modal: true,
-  title: '',
-  autoOpen: false,
-  buttons:
-    [
-      {
-        text: 'Abbrechen',
-        click: function () {
-          $('#fs-profile-rate-comment').dialog('close')
-        }
-      },
-      {
-        text: 'Absenden',
-        click: function () {
-          ajreq('rate', {
-            app: 'profile',
-            type: 2,
-            id: $('#profile-rate-id').val(),
-            message: $('#fsprofileratemsg').val()
-          })
-        }
-      }
-    ]
-}).siblings('.ui-dialog-titlebar').remove()
 
 if (serverData.user.may) {
   socket.connect()