+++ src/Modules/Profile/Profile.js
@@ -1,4 +1,3 @@
 import '@/core'
 import '@/globals'
 import './Profile.css'
-import 'magnific-popup'