+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Move master-update function for regions to the rest controller !1547 @alex.simm 
 - Add missing endpoint for deleting forum threads !1545 #913 @alex.simm
 - Use rest endpoints for basket deletion and the coordinates on the baskets map !1550 @alex.simm 
+- Remove Magnific Popup by rewriting trust banana UI code to fancybox !1530 @ChrisOelmueller
 
 ## Dev/Test/CI stuff